+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/arguments/ShopkeeperByIdArgument.java
@@ -1,6 +1,5 @@
 package com.nisovin.shopkeepers.commands.arguments;
 
-import java.util.Collections;
 
 import com.nisovin.shopkeepers.api.ShopkeepersAPI;
 import com.nisovin.shopkeepers.api.shopkeeper.Shopkeeper;
@@ -41,12 +40,8 @@ public class ShopkeeperByIdArgument extends ObjectByIdArgument<Integer, Shopkeep
 	}
 
 	@Override
-	public Text getInvalidArgumentErrorMsg(String argumentInput) {
-		if (argumentInput == null) argumentInput = "";
-		Text text = Messages.commandShopkeeperArgumentInvalid;
-		text.setPlaceholderArguments(this.getDefaultErrorMsgArgs());
-		text.setPlaceholderArguments(Collections.singletonMap("argument", argumentInput));
-		return text;
+	protected Text getInvalidArgumentErrorMsgText() {
+		return Messages.commandShopkeeperArgumentInvalid;
 	}
 
 	@Override