+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/arguments/ShopkeeperByNameArgument.java
@@ -1,6 +1,5 @@
 package com.nisovin.shopkeepers.commands.arguments;
 
-import java.util.Collections;
 import java.util.Optional;
 import java.util.stream.Stream;
 
@@ -45,12 +44,8 @@ public class ShopkeeperByNameArgument extends ObjectByIdArgument<String, Shopkee
 	}
 
 	@Override
-	public Text getInvalidArgumentErrorMsg(String argumentInput) {
-		if (argumentInput == null) argumentInput = "";
-		Text text = Messages.commandShopkeeperArgumentInvalid;
-		text.setPlaceholderArguments(this.getDefaultErrorMsgArgs());
-		text.setPlaceholderArguments(Collections.singletonMap("argument", argumentInput));
-		return text;
+	protected Text getInvalidArgumentErrorMsgText() {
+		return Messages.commandShopkeeperArgumentInvalid;
 	}
 
 	@Override