+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/lib/arguments/PlayerByUUIDArgument.java
@@ -1,6 +1,5 @@
 package com.nisovin.shopkeepers.commands.lib.arguments;
 
-import java.util.Collections;
 import java.util.UUID;
 
 import org.bukkit.Bukkit;
@@ -41,12 +40,8 @@ public class PlayerByUUIDArgument extends ObjectByIdArgument<UUID, Player> {
 	}
 
 	@Override
-	public Text getInvalidArgumentErrorMsg(String argumentInput) {
-		if (argumentInput == null) argumentInput = "";
-		Text text = Messages.commandPlayerArgumentInvalid;
-		text.setPlaceholderArguments(this.getDefaultErrorMsgArgs());
-		text.setPlaceholderArguments(Collections.singletonMap("argument", argumentInput));
-		return text;
+	protected Text getInvalidArgumentErrorMsgText() {
+		return Messages.commandPlayerArgumentInvalid;
 	}
 
 	@Override